Add tests to exercise current orientation behavior
To stay backwards compatible and ensure that current users are not
affected by upcoming changes, we need to test the current behavior of
the `flip`, `flop`, and `rotate` transforms when applied to images with
EXIF orientation metadata.

The images added to __fixtures__ are all rotated via EXIF orientation.
Whether or not they appear correctly in your image viewer is dependent
on its ability to read that orientation metadata. The resulting PNG
files show the problems that people run into when they don't account for
orientation metadata in their image processing. The next commit will
introduce a new setting that attempts to give developers the results
they actually expect.