Add experimental setting: preserveInitialOrientation
This experimental setting is intended to give developers an easy way to
treat their images intuitively based on how they actually look on
screen, instead of how they are actually written on disk.

There is an overhead to certain orientations that might be too high a
cost to pay, but that is why this is experimental. Working on this
feature has made me believe that the best place to add the DX
improvement is actually within Sharp itself. In the meantime there is
something here worth trying out.